fix http status code usage

+++ src/NuGetGallery/Controllers/OrganizationsController.cs
@@ -101,12 +101,12 @@ namespace NuGetGallery
                 || ActionsRequiringPermissions.ManageAccount.CheckPermissions(GetCurrentUser(), account)
                     != PermissionsCheckResult.Allowed)
             {
-                return Json((int)HttpStatusCode.Forbidden, Strings.Unauthorized);
+                return Json(HttpStatusCode.Forbidden, Strings.Unauthorized);
             }
 
             if (!account.Confirmed)
             {
-                return Json((int)HttpStatusCode.BadRequest, Strings.Member_OrganizationUnconfirmed);
+                return Json(HttpStatusCode.BadRequest, Strings.Member_OrganizationUnconfirmed);
             }
 
             try
@@ -126,7 +126,7 @@ namespace NuGetGallery
             }
             catch (EntityException e)
             {
-                return Json((int)HttpStatusCode.BadRequest, e.Message);
+                return Json(HttpStatusCode.BadRequest, e.Message);
             }
         }
 
@@ -200,7 +200,7 @@ namespace NuGetGallery
                 || ActionsRequiringPermissions.ManageAccount.CheckPermissions(GetCurrentUser(), account)
                     != PermissionsCheckResult.Allowed)
             {
-                return Json((int)HttpStatusCode.Forbidden, Strings.Unauthorized);
+                return Json(HttpStatusCode.Forbidden, Strings.Unauthorized);
             }
 
             try
@@ -211,7 +211,7 @@ namespace NuGetGallery
             }
             catch (EntityException e)
             {
-                return Json((int)HttpStatusCode.BadRequest, e.Message);
+                return Json(HttpStatusCode.BadRequest, e.Message);
             }
         }
 
@@ -226,12 +226,12 @@ namespace NuGetGallery
                 || ActionsRequiringPermissions.ManageAccount.CheckPermissions(GetCurrentUser(), account)
                     != PermissionsCheckResult.Allowed)
             {
-                return Json((int)HttpStatusCode.Forbidden, Strings.Unauthorized);
+                return Json(HttpStatusCode.Forbidden, Strings.Unauthorized);
             }
 
             if (!account.Confirmed)
             {
-                return Json((int)HttpStatusCode.BadRequest, Strings.Member_OrganizationUnconfirmed);
+                return Json(HttpStatusCode.BadRequest, Strings.Member_OrganizationUnconfirmed);
             }
 
             try
@@ -243,7 +243,7 @@ namespace NuGetGallery
             }
             catch (EntityException e)
             {
-                return Json((int)HttpStatusCode.BadRequest, e.Message);
+                return Json(HttpStatusCode.BadRequest, e.Message);
             }
         }
 
@@ -261,12 +261,12 @@ namespace NuGetGallery
                 ActionsRequiringPermissions.ManageAccount.CheckPermissions(currentUser, account)
                     != PermissionsCheckResult.Allowed))
             {
-                return Json((int)HttpStatusCode.Forbidden, Strings.Unauthorized);
+                return Json(HttpStatusCode.Forbidden, Strings.Unauthorized);
             }
 
             if (!account.Confirmed)
             {
-                return Json((int)HttpStatusCode.BadRequest, Strings.Member_OrganizationUnconfirmed);
+                return Json(HttpStatusCode.BadRequest, Strings.Member_OrganizationUnconfirmed);
             }
 
             try
@@ -277,7 +277,7 @@ namespace NuGetGallery
             }
             catch (EntityException e)
             {
-                return Json((int)HttpStatusCode.BadRequest, e.Message);
+                return Json(HttpStatusCode.BadRequest, e.Message);
             }
         }
